requirements.txt
python-binance==0.6.6